FuncionPago
Funcionalidad boton de pago y Almacenamiento de encabezado y dellates de la factura en BD

+++ PAWS_ProyectoFinal/Models/ItemCarrito.cs
@@ -5,8 +5,8 @@
 		public int ProductoId { get; set; }
 		public string NomProducto { get; set; }
 		public int Cantidad { get; set; }
-		public decimal Precio { get; set; }
-		public decimal Total
+		public int Precio { get; set; }
+		public int Total
 		{
 			get { return Cantidad * Precio; }
 		}
